1. 添加显示介绍配置 luaide.isShowDest 默认为false 只显示一次,如需重复显示修改为true 	2. 修改代码格式化 换行处理 和 " 转义 bug 	3. 修改代码提示  function  方法中 定义的local 变量 无法提示 二级变量 的bug 	4. 添加数据统计接口 统计在线人数, 如果有反感这一行为的请联系我,后期考虑添加配置 	5. 优化debug  将 lua 和luajit 调试文件进行分离 coocs 和unity 如果使用luajit 的调试文件请使用luaDebugjit.lua 文件进行调试

+++ src/luatool/LuaParse.ts
@@ -139,8 +139,7 @@ export class LuaParse {
       token.index = tokens.length;
       tokens.push(token);
     }
-    var date1:Date = new Date();
-    var time:number = date1.getMilliseconds() - data.getMilliseconds()
+   
     this.tokens = tokens;
     this.luaInfoManager.init(this, this.currentUri);
     this.tokenIndex = 0;
@@ -161,23 +160,13 @@ export class LuaParse {
             break;
         }
       }
-      // console.log("正确了")
+      
       //正确了删除错误提示
+      if(this.diagnosticCollection  && this.diagnosticCollection.has(this.currentUri)){
       this.diagnosticCollection.delete(this.currentUri)
+      }
       this.luaInfoManager.getFcim(this.currentUri).selfToGolbal();
     } else {
-      // var isAdd:boolean = true
-      // this.errorFilePaths.forEach(uri=>{
-      //   if(this.currentUri.path == uri.path)
-      //   {
-      //     isAdd = false;
-      //   }
-      // })
-      // if(isAdd)
-      // {
-      //   this.errorFilePaths.push(this.currentUri);
-      // }
-       this.luaInfoManager.getFcim(this.currentUri).selfToGolbal();
       console.log("错误的")
     }
